feat: ftype box parsing
Support for ftype box parsing, including major brands, minor
versions, and compatible brands.

+++ src/reader/MediaFile.h
@@ -1,4 +1,3 @@
-
 #pragma once
 
 #include <string>
@@ -6,11 +5,11 @@
 class MediaFile {
    public:
     int id();
-    uint32_t size();
-    uint32_t buffer_index();
-    uint8_t u8_from(uint32_t pos);
-    uint16_t u16_from(uint32_t pos);
-    uint32_t u32_from(uint32_t pos);
+    size_t size();
+    size_t buffer_index();
+    uint8_t u8_from(size_t pos);
+    uint16_t u16_from(size_t pos);
+    uint32_t u32_from(size_t pos);
     std::string name();
     std::string status();
     bool is_invalid();
@@ -28,8 +27,8 @@ class MediaFile {
     int _id;
     bool _is_closed;
     bool _is_invalid;
-    uint32_t _size;
-    uint32_t _fill_pos;
+    size_t _size;
+    size_t _fill_pos;
     std::string _name;
 
     uint8_t* _buffer;
